+++ src/main/java/io/github/mjaroslav/sharedjava/function/ToIntTerFunction.java
@@ -1,0 +1,6 @@
+package io.github.mjaroslav.sharedjava.function;
+
+@FunctionalInterface
+public interface ToIntTerFunction<T, U, V> {
+    int applyAsInt(T t, U u, V v);
+}