+++ src/main/java/io/github/mjaroslav/sharedjava/tuple/pair/MonoPair.java
@@ -1,0 +1,4 @@
+package io.github.mjaroslav.sharedjava.tuple.pair;
+
+public interface MonoPair<T> extends Pair<T, T> {
+}