+++ src/main/java/io/github/mjaroslav/sharedjava/tuple/trio/MonoTrio.java
@@ -1,0 +1,6 @@
+package io.github.mjaroslav.sharedjava.tuple.trio;
+
+import io.github.mjaroslav.sharedjava.tuple.pair.MonoPair;
+
+public interface MonoTrio<T> extends Trio<T, T, T>, MonoPair<T> {
+}