+++ src/main/java/io/github/mjaroslav/sharedjava/tuple/unit/NumberUnit.java
@@ -1,0 +1,10 @@
+package io.github.mjaroslav.sharedjava.tuple.unit;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class NumberUnit<T extends Number> extends SimpleUnit<T> {
+    public NumberUnit(T x) {
+        setX(x);
+    }
+}