+++ CHANGELOG.md
@@ -14,5 +14,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Bits](src/main/java/io/github/mjaroslav/sharedjava/format/Bits.java) utility class.
 - [LazySupplier](src/main/java/io/github/mjaroslav/sharedjava/function/LazySupplier.java) realization
   for `java.util.function.Supplier`.
+- [Unit](src/main/java/io/github/mjaroslav/sharedjava/tuple/unit/Unit.java) tuple with simple and number realization.
+- [Pair](src/main/java/io/github/mjaroslav/sharedjava/tuple/pair/Pair.java) tuple with simple, mono and number
+  realization.
+- [Trio](src/main/java/io/github/mjaroslav/sharedjava/tuple/trio/Trio.java) tuple with simple, mono and number
+  realization.
+- [Tuples](src/main/java/io/github/mjaroslav/sharedjava/tuple/Tuples.java) helper for object creating (like as
+  Collections class in Java).
+- [TerFunction](src/main/java/io/github/mjaroslav/sharedjava/function/TerFunction.java) just a function with 3
+  arguments.
+- [QuaterFunction](src/main/java/io/github/mjaroslav/sharedjava/function/QuaterFunction.java) just a function with 4
+  arguments.
+- [TerPredicate](src/main/java/io/github/mjaroslav/sharedjava/function/TerPredicate.java) just a predicate with 3
+  arguments.
+- [QuaterPredicate](src/main/java/io/github/mjaroslav/sharedjava/function/QuaterPredicate.java) just a predicate with 4
+  arguments.
+- [ToIntTerFunction](src/main/java/io/github/mjaroslav/sharedjava/function/ToIntTerFunction.java) just int function with
+  3 arguments.
+- [ToIntQuaterFunction](src/main/java/io/github/mjaroslav/sharedjava/function/ToIntQuaterFunction.java) just int
+  function with 4 arguments.
 
 [unreleased]: https://github.com/MJaroslav/Shared-Java/